requirements.txt
sanic
python-socketio
sanic-jinja2